+++ ta4j-core/src/main/java/org/ta4j/core/indicators/numeric/BinaryOperation.java
@@ -27,15 +27,15 @@ import java.util.function.BinaryOperator;
 
 import org.ta4j.core.BarSeries;
 import org.ta4j.core.Indicator;
+import org.ta4j.core.indicators.helpers.ConstantIndicator;
 import org.ta4j.core.num.Num;
 
 /**
  * Objects of this class defer evaluation of an arithmetic operation.
  *
  * <p>
- * This is a lightweight version of the
- * {@link org.ta4j.core.indicators.helpers.CombineIndicator CombineIndicator};
- * it doesn't cache.
+ * This is a lightweight, non-cached implementation for binary operations
+ * between two indicators.
  */
 public class BinaryOperation implements Indicator<Num> {
 
@@ -47,7 +47,7 @@ public class BinaryOperation implements Indicator<Num> {
      * @return {@code left + right}, rounded as necessary
      * @see Num#plus
      */
-    public static BinaryOperation sum(Indicator<Num> left, Indicator<Num> right) {
+    public static BinaryOperation sum(final Indicator<Num> left, final Indicator<Num> right) {
         return new BinaryOperation(Num::plus, left, right);
     }
 
@@ -59,7 +59,7 @@ public class BinaryOperation implements Indicator<Num> {
      * @return {@code left - right}, rounded as necessary
      * @see Num#minus
      */
-    public static BinaryOperation difference(Indicator<Num> left, Indicator<Num> right) {
+    public static BinaryOperation difference(final Indicator<Num> left, final Indicator<Num> right) {
         return new BinaryOperation(Num::minus, left, right);
     }
 
@@ -71,7 +71,7 @@ public class BinaryOperation implements Indicator<Num> {
      * @return {@code left * right}, rounded as necessary
      * @see Num#multipliedBy
      */
-    public static BinaryOperation product(Indicator<Num> left, Indicator<Num> right) {
+    public static BinaryOperation product(final Indicator<Num> left, final Indicator<Num> right) {
         return new BinaryOperation(Num::multipliedBy, left, right);
     }
 
@@ -83,7 +83,7 @@ public class BinaryOperation implements Indicator<Num> {
      * @return {@code left / right}, rounded as necessary
      * @see Num#dividedBy
      */
-    public static BinaryOperation quotient(Indicator<Num> left, Indicator<Num> right) {
+    public static BinaryOperation quotient(final Indicator<Num> left, final Indicator<Num> right) {
         return new BinaryOperation(Num::dividedBy, left, right);
     }
 
@@ -97,7 +97,7 @@ public class BinaryOperation implements Indicator<Num> {
      *         {@code right}. If they are equal, {@code left} is returned.
      * @see Num#min
      */
-    public static BinaryOperation min(Indicator<Num> left, Indicator<Num> right) {
+    public static BinaryOperation min(final Indicator<Num> left, final Indicator<Num> right) {
         return new BinaryOperation(Num::min, left, right);
     }
 
@@ -111,30 +111,118 @@ public class BinaryOperation implements Indicator<Num> {
      *         {@code right}. If they are equal, {@code left} is returned.
      * @see Num#max
      */
-    public static BinaryOperation max(Indicator<Num> left, Indicator<Num> right) {
+    public static BinaryOperation max(final Indicator<Num> left, final Indicator<Num> right) {
         return new BinaryOperation(Num::max, left, right);
+    }
+
+    // Overloaded methods for operations with constants
+
+    /**
+     * Returns an {@code Indicator} whose value is {@code (indicator + addend)}.
+     *
+     * @param indicator the indicator
+     * @param addend    the coefficient to add
+     * @return {@code indicator + addend}, rounded as necessary
+     */
+    public static BinaryOperation sum(final Indicator<Num> indicator, final Number addend) {
+        final var constantIndicator = new ConstantIndicator<>(indicator.getBarSeries(),
+                indicator.getBarSeries().numFactory().numOf(addend));
+        return new BinaryOperation(Num::plus, indicator, constantIndicator);
+    }
+
+    /**
+     * Returns an {@code Indicator} whose value is {@code (indicator - subtrahend)}.
+     *
+     * @param indicator  the indicator
+     * @param subtrahend the subtrahend to subtract
+     * @return {@code indicator - subtrahend}, rounded as necessary
+     */
+    public static BinaryOperation difference(final Indicator<Num> indicator, final Number subtrahend) {
+        final var constantIndicator = new ConstantIndicator<>(indicator.getBarSeries(),
+                indicator.getBarSeries().numFactory().numOf(subtrahend));
+        return new BinaryOperation(Num::minus, indicator, constantIndicator);
+    }
+
+    /**
+     * Returns an {@code Indicator} whose value is
+     * {@code (indicator * coefficient)}.
+     *
+     * @param indicator   the indicator
+     * @param coefficient the coefficient to multiply by
+     * @return {@code indicator * coefficient}, rounded as necessary
+     */
+    public static BinaryOperation product(final Indicator<Num> indicator, final Number coefficient) {
+        final var constantIndicator = new ConstantIndicator<>(indicator.getBarSeries(),
+                indicator.getBarSeries().numFactory().numOf(coefficient));
+        return new BinaryOperation(Num::multipliedBy, indicator, constantIndicator);
+    }
+
+    /**
+     * Returns an {@code Indicator} whose value is
+     * {@code (indicator / coefficient)}.
+     *
+     * @param indicator   the indicator
+     * @param coefficient the coefficient to divide by
+     * @return {@code indicator / coefficient}, rounded as necessary
+     */
+    public static BinaryOperation quotient(final Indicator<Num> indicator, final Number coefficient) {
+        final var constantIndicator = new ConstantIndicator<>(indicator.getBarSeries(),
+                indicator.getBarSeries().numFactory().numOf(coefficient));
+        return new BinaryOperation(Num::dividedBy, indicator, constantIndicator);
+    }
+
+    /**
+     * Returns the minimum of {@code indicator} and {@code constant} as an
+     * {@code Indicator}.
+     *
+     * @param indicator the indicator
+     * @param constant  the constant to compare with
+     * @return the {@code Indicator} whose value is the smaller of {@code indicator}
+     *         and {@code constant}. If they are equal, {@code indicator} is
+     *         returned.
+     */
+    public static BinaryOperation min(final Indicator<Num> indicator, final Number constant) {
+        final var constantIndicator = new ConstantIndicator<>(indicator.getBarSeries(),
+                indicator.getBarSeries().numFactory().numOf(constant));
+        return new BinaryOperation(Num::min, indicator, constantIndicator);
+    }
+
+    /**
+     * Returns the maximum of {@code indicator} and {@code constant} as an
+     * {@code Indicator}.
+     *
+     * @param indicator the indicator
+     * @param constant  the constant to compare with
+     * @return the {@code Indicator} whose value is the greater of {@code indicator}
+     *         and {@code constant}. If they are equal, {@code indicator} is
+     *         returned.
+     */
+    public static BinaryOperation max(final Indicator<Num> indicator, final Number constant) {
+        final var constantIndicator = new ConstantIndicator<>(indicator.getBarSeries(),
+                indicator.getBarSeries().numFactory().numOf(constant));
+        return new BinaryOperation(Num::max, indicator, constantIndicator);
     }
 
     private final BinaryOperator<Num> operator;
     private final Indicator<Num> left;
     private final Indicator<Num> right;
 
-    private BinaryOperation(BinaryOperator<Num> operator, Indicator<Num> left, Indicator<Num> right) {
+    private BinaryOperation(final BinaryOperator<Num> operator, final Indicator<Num> left, final Indicator<Num> right) {
         this.operator = operator;
         this.left = left;
         this.right = right;
     }
 
     @Override
-    public Num getValue(int index) {
-        Num n1 = left.getValue(index);
-        Num n2 = right.getValue(index);
+    public Num getValue(final int index) {
+        final var n1 = left.getValue(index);
+        final var n2 = right.getValue(index);
         return operator.apply(n1, n2);
     }
 
     @Override
     public int getCountOfUnstableBars() {
-        return 0;
+        return Math.max(left.getCountOfUnstableBars(), right.getCountOfUnstableBars());
     }
 
     @Override